requirements.txt
python-telegram-bot==20.7
requests==2.31.0
matplotlib==3.8.0
Pillow==10.1.0